+++ src/core/application-config.ts
@@ -1,5 +1,9 @@
+import { IoAdapter } from '@nestjs/websockets/adapters/io-adapter';
+import { SocketIoAdapter } from '@nestjs/common/interfaces';
+
 export class ApplicationConfig {
     private globalPrefix = '';
+    private ioAdapter: SocketIoAdapter = IoAdapter as any;
 
     public setGlobalPrefix(prefix: string) {
         this.globalPrefix = prefix;
@@ -7,5 +11,13 @@ export class ApplicationConfig {
 
     public getGlobalPrefix() {
         return this.globalPrefix;
+    }
+
+    public setIoAdapter(ioAdapter: SocketIoAdapter) {
+        this.ioAdapter = ioAdapter;
+    }
+
+    public getIoAdapter(): SocketIoAdapter {
+        return this.ioAdapter;
     }
 }